+++ backend/src/resolvers/user.resolver.ts
@@ -1,28 +1,43 @@
-import { Arg, Float, Ctx,  Mutation, Query, Resolver, Authorized } from "type-graphql";
+import { Arg, Ctx, Mutation, Query, Resolver, Authorized } from "type-graphql";
 import UsersService from "../services/users.service";
-import { User, CreateUserInput, UpdateUserInput, ROLE, Message, InputLogin } from "../entities/user.entity";
+import {
+  User,
+  CreateUserInput,
+  UpdateUserInput,
+  ROLE,
+  Message,
+  InputLogin,
+} from "../entities/user.entity";
 import * as argon2 from "argon2";
 import { SignJWT } from "jose";
 import { MyContext } from "..";
 import Cookies from "cookies";
 import { Project } from "../entities/project.entity";
-import { CreateUserProjectAccessesInput } from "../entities/userProjectAccesses.entity";
 
 @Resolver()
 export class UserResolver {
 
-  @Authorized(['ADMIN'])
+  @Authorized(["ADMIN"])
   @Query(() => [User])
   async listUsers() {
     const users = await new UsersService().list();
     return users;
   }
 
+  @Authorized()
   @Query(() => [User])
   async listUsersByRole(@Arg("role") role: ROLE) {
     const users = await new UsersService().listByRole(role);
     return users;
   }
+
+  @Authorized()
+  @Query(() => [User])
+  async listUsersByPseudo(@Arg("pseudo") pseudo: string) {
+    const users = await new UsersService().listUsersByPseudo(pseudo);
+    return users;
+  }
+
 
   @Query(() => User)
   async findUserById(@Arg("id") id: string) {
@@ -31,6 +46,7 @@ export class UserResolver {
     if (!userById) throw new Error("Please note, the client does not exist");
     return userById;
   }
+
 
   @Query(() => User)
   async findUserByEmail(@Arg("email") email: string) {
@@ -42,7 +58,8 @@ export class UserResolver {
   @Query(() => User)
   async findUserByPseudo(@Arg("pseudo") pseudo: string) {
     const userByPseudo = await new UsersService().findByPseudo(pseudo);
-    if (!userByPseudo) throw new Error("Please note, the client does not exist");
+    if (!userByPseudo)
+      throw new Error("Please note, the client does not exist");
     return userByPseudo;
   }
 
@@ -55,7 +72,6 @@ export class UserResolver {
       user = await new UsersService().findByEmail(infos.email);
     } else {
       user = await new UsersService().findByPseudo(infos.pseudo);
-
     }
     if (!user) {
       throw new Error("Check your login information !");
@@ -64,11 +80,11 @@ export class UserResolver {
     const isPasswordValid = await argon2.verify(user.password, infos.password);
     const m = new Message();
     if (isPasswordValid) {
-      const token = await new SignJWT({ 
+      const token = await new SignJWT({
         email: user.email,
         role: user.role,
         pseudo: user.pseudo,
-        id : user.id
+        id: user.id,
       })
         .setProtectedHeader({ alg: "HS256", typ: "jwt" })
         .setExpirationTime("2h")
@@ -77,7 +93,7 @@ export class UserResolver {
       let cookies = new Cookies(ctx.req, ctx.res);
       cookies.set("token", token, { httpOnly: true });
 
-      m.message = "Welcome !";
+      m.message = "You are connected!";
       m.success = true;
     } else {
       m.message = "Check your login information !";
@@ -85,31 +101,51 @@ export class UserResolver {
     }
     return m;
   }
-  
+
   @Mutation(() => User)
   async register(@Arg("data") data: CreateUserInput) {
-    const user = await new UsersService().findByEmail(data.email);
+    const email = await new UsersService().findByEmail(data.email);
     const pseudo = await new UsersService().findByPseudo(data.pseudo);
 
-    if (user && pseudo) {
+    if (email && pseudo) {
       throw new Error("This email and pseudo is already in use!");
-    } else if (user) {
+    } else if (email) {
       throw new Error("This email is already in use!");
     } else if (pseudo) {
       throw new Error("This pseudo is already in use!");
     }
-    
+
     const newUser = await new UsersService().create(data);
     return newUser;
   }
 
   @Authorized()
   @Mutation(() => Message)
-  async updateUser(@Arg("data") data: UpdateUserInput) {
+  async updateUser(@Arg("data") data: UpdateUserInput, @Ctx() ctx: MyContext) {
+
+    if (!ctx.user)
+      throw new Error("Access denied! You need to be authenticated to perform this action!");
+
+    if (ctx.user.role !== "ADMIN" && data.id != ctx.user.id)
+      throw new Error("You must be a site administrator to do this action!"); 
+
     const { id, ...otherData } = data;
     if (otherData.password) {
       otherData.password = await argon2.hash(otherData.password);
     }
+
+    if (data?.email) {
+      const checkUserEmail = await new UsersService().findByEmail(data?.email);
+      if (checkUserEmail)
+        throw new Error("This email already exists in our database!");
+    }
+
+    if (data?.pseudo) {
+      const checkUserPseudo = await new UsersService().findByEmail(data?.email);
+      if (checkUserPseudo)
+        throw new Error("This pseudo already exists in our database!");
+    }
+
     const updateUser = await new UsersService().update(+id, otherData);
     const m = new Message();
     if (updateUser) {
@@ -121,13 +157,20 @@ export class UserResolver {
     }
     return m;
   }
-  
+
   @Authorized()
   @Mutation(() => Message)
-  async deleteUser(@Arg("id") id: number) {
+  async deleteUser(@Arg("id") id: number, @Ctx() ctx: MyContext) {
+
+    if (!ctx.user)
+      throw new Error("Access denied! You need to be authenticated to perform this action!");
+
+    if (ctx.user.role !== "ADMIN" && id != ctx.user.id)
+      throw new Error("You must be a site administrator to do this action!"); 
+
     const delUser = await new UsersService().delete(id);
     const m = new Message();
-    
+
     if (delUser) {
       m.message = "User deleted!";
       m.success = true;
@@ -135,10 +178,10 @@ export class UserResolver {
       m.message = "Unable to delete user!";
       m.success = false;
     }
-    
+
     return m;
   }
-  
+
   @Authorized()
   @Query(() => Message)
   async logout(@Ctx() ctx: MyContext) {
@@ -149,14 +192,13 @@ export class UserResolver {
     const m = new Message();
     m.message = "You have been disconnected !";
     m.success = true;
-    
+
     return m;
   }
 
-  @Authorized()
   @Query(() => [Project])
-  async listLikeProject(@Arg("userId") userId: number) {
-    const projects = await new UsersService().listLikedProjects(userId);
+  async listLikeProject(@Arg("userId") userId: string) {
+    const projects = await new UsersService().listLikedProjects(+userId);
     if (projects.length === 0) {
       throw new Error("You have no plans !");
     }
@@ -165,9 +207,14 @@ export class UserResolver {
 
   @Authorized()
   @Mutation(() => Message)
-  async addLikeProject(@Arg("userId") userId: number, @Arg("projectId") projectId: number) {
-
-    const likedProjects = await new UsersService().likeProject(userId, projectId);
+  async addLikeProject(
+    @Arg("userId") userId: number,
+    @Arg("projectId") projectId: number
+  ) {
+    const likedProjects = await new UsersService().likeProject(
+      userId,
+      projectId
+    );
 
     const m = new Message();
 
@@ -180,14 +227,18 @@ export class UserResolver {
     }
 
     return m;
-
   }
 
   @Authorized()
   @Mutation(() => Message)
-  async deleteLikeProject(@Arg("userId") userId: number, @Arg("projectId") projectId: number) {
-
-    const likedProjects = await new UsersService().dislikeProject(userId, projectId);
+  async deleteLikeProject(
+    @Arg("userId") userId: number,
+    @Arg("projectId") projectId: number
+  ) {
+    const likedProjects = await new UsersService().dislikeProject(
+      userId,
+      projectId
+    );
 
     const m = new Message();
 
@@ -200,59 +251,11 @@ export class UserResolver {
     }
 
     return m;
-    
   }
 
-  @Authorized()
-  @Query(() => [Project])
-  async listAccesProject (@Arg("userId") userId: number) {
-    const listAccesProject = await new UsersService().findUsersByAccessesProject(userId);
-    return listAccesProject;
+  @Query(() => User)
+  async findProjectOwner(@Arg("projectId") projectId: string) {
+    const projectOwner = await new UsersService().findOwner(+projectId);
+    return projectOwner;
   }
-
-  @Authorized()
-  @Mutation(() => Message)
-  async addAccessProject (@Arg("data") data: CreateUserProjectAccessesInput) {
-    const user = await new UsersService().findByAccessesProject(data.user_id, data.project_id);
-    
-    if (user) {
-      throw new Error("This user already has access to this project!");
-    }
-
-    if (user) throw new Error("This name of project is already in use!");
-    
-    const newUserAccessesProject = await new UsersService().createAccessesProject(data);
-    
-    const m = new Message();
-
-    if (newUserAccessesProject) {
-      m.message = "Add user project!";
-      m.success = true;
-    } else {
-      m.message = "Unable to add user project!";
-      m.success = false;
-    }
-
-    return m;
-  }
-
-  @Authorized()
-  @Mutation(() => Message)
-  async deleteAccessProject (@Arg("userId") userId: number, @Arg("projectId") projectId: number) {
-
-    const deleteUserAccessesProject = await new UsersService().deleteAccessesProject(userId, projectId);
-    
-    const m = new Message();
-
-    if (deleteUserAccessesProject) {
-      m.message = "Delete user project!";
-      m.success = true;
-    } else {
-      m.message = "Unable to delete user project!";
-      m.success = false;
-    }
-
-    return m;
-  }
-
 }
